pyproject.toml
[project]
name = "parakeet-dictation"
version = "0.2.0"
description = "A dictation tool powered by Parakeet"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.12,<3.13"
dependencies = [
    "parakeet-mlx",
    "numpy<2.3",
    "pyaudio~=0.2.14",
    "rumps~=0.4.0",
    "pyperclip~=1.9.0",
    "python-dotenv~=1.1.1",
    "pyobjc-framework-cocoa~=11.1; platform_system == 'Darwin'",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project.scripts]
parakeet-dictation = "parakeet_dictation.main:main"

[tool.setuptools.packages.find]
where = ["src"]

[project.optional-dependencies]
dev = ["build", "pytest", "ruff", "mypy", "py2app"]

[tool.pytest.ini_options]
testpaths = ["tests"]
norecursedirs = ["build", "dist", "build.bak.*", "dist.bak.*"]
